feat: add navbar to all pages
