+++ services/admin-client/pages/_app.tsx
@@ -1,0 +1,12 @@
+// import '../styles/globals.css';
+import type { AppProps } from 'next/app';
+import Layout from '../components/Layout';
+
+function MyApp({ Component, pageProps }: AppProps) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
+}
+export default MyApp;
